seafile: Depends on EOL and CVE-ridden Django 1.5, therefore abandon
Django 1.5 is EOL and has a number of CVE's.  PIL also has a major CVE.
seafile-seahub (without which there is not point to the other seafile packages)
depends on Django 1.5, and, through Djblets 0.6 (old djblets is required because
of old Django) depends on PIL.

Therefore until seafile gets updated to newer Django, move seafile to abandoned

We also abandon deprecated libevhtp-1.1 for which seafile is the only consumer, as
well as libsearpc which is seafile-specific library.