netatalk: move package here from packages feed
It's time to park this package.

AppleShare products have been unused for a while now (since Mac OS 9.2.2)
around 2002.
So, there should be fewer users requiring this package.

Last update of netatalk was in December 2018. Not sure if newer updates
will be created.

It's time to cut the cord on our end and move it to the abandoned packages.

Info: https://en.wikipedia.org/wiki/AppleShar

Signed-off-by: Alexandru Ardelean <ardeleanalex@gmail.com>